Further additions to SSH public key support.
Added SSH key file field in Preferences - this will be used by Paramiko
as well as searching `~/.ssh`.

The password field is ignored if an SSH key file is specified, unless
that key file does not correspond with the server, in which case
password authentication will be attempted.

I tried using the password field to unlock a locked private key, but
with no luck; so that's an area for improvement in the future.

If the SSH key file field is left empty, Paramiko will still look in
`~/.ssh`, so it is purely used for specifying a key file outside that
directory.